pi_floppy/requirements.txt
Werkzeug==2.2.2
Flask==2.2.3
flask-socketio==5.3.2
python-socketio==5.7.2
requests
gevent==22.10.2
gevent-websocket
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-pn532
